Fix more (mostly mob) harddels

+++ code/modules/mob/mob.dm
@@ -796,7 +796,8 @@ GLOBAL_VAR_INIT(exploit_warn_spam_prevention, 0)
 		var/obj/effect/proc_holder/spell/S = X
 		if(istype(S, spell))
 			mob_spell_list -= S
-			qdel(S)
+			if(!QDELETED(S))
+				qdel(S)
 	if(client)
 		client << output(null, "statbrowser:check_spells")
